QQuickMenu: guard agains crashing, if the visual parent is set to null
The application can do silly things as setting a menus
visual parent to null before showing it. This will currently
cause a crash. And doing so will also make the menu not
showing when using non-native menus, since the menu will
not be a part of the scene.

This patch will there check if a menus visual parent is set
before trying to show it.

Change-Id: I6094b3ddee88189db59a82c6791fe90456c8da92
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>