Adjust locale correctly in tst_spinbox
The test for displayTextChangedSpy.count is flaky and failed
repeatedly in CI because the signal is emitted once more than
expected. The reason was that setting locale changed the displayText
depending on the native locale on the testing platform.
Further down in the test, the validator failed because the locale was
not set correctly on the validator.

This patch sets locale in the correct place and also improves the
testing.

Fixes: QTBUG-114694
Pick-to: 6.6
Change-Id: Ied4b41fbee5e25fabceabda6815272c44aa1f832
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit cf15ba9b085ac783fd3315f1b103c012947cb7db)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>