Handle more potential underlying types
Since f3bcbfd6a50fb7e74f4ff6714d3b3066fa74e253, we use the underlying
type of an enum when the binding is detected as a resolved enum and
the property is of enum type.

When using an enums underlying type, we can receive types that were
formerly not possible, because we would normalize them to plain int.

Fix this by teaching the object creator to handle assignments of
signed and unsigned integer types.

Fixes: QTBUG-123428
Change-Id: I3635b7de048e0c6ddf79381679aafee89a5b3def
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>