qmltc: Avoid befriending the root component twice
When `qmltc` compiles a type to C++, if the type is not the root of the
document, it will generate code to befriend the root component.

Similarly, it will generate code to befriend the direct parent of the
component.

In both cases this is done as the component may be created from those
other generated types by accessing some private members of the generated
API.

When a component is not the root of the document, is not an inline
component and its parent is the root of the document, `qmltc` will
generate the same friend declaration twice, once to befriend the root of
the document and once to befriend the direct parent of the type.

For example, if we have some component `Foo` and a root component that
refers to it, say:

```
Window {
    Foo {}
}
```

The generated code for `Foo` will befriend the generated type for the
root component twice.

While multiple friend declaration are not an issue, they are generally
uncommon, as there is no reason to befriend something twice, so that
they may generally stand out as an error or a typo, while making the
code noisier.

Thus, to avoid befriending the document root twice for generated
non-root, non-inline components whose parent is the root of the
document, the code that generates the friend declaration for the direct
parent of such a component was conditioned on that parent not being the
document root, as we would have already befriended it at that point of
the compilation.

Change-Id: Ie6eb4b786317c3c555718b594c689c085b4f2336
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>