qmltyperegistrar: Clarify confusion about isCreatable
For value types isCreatable means that the type is constructible. For
object types it means that it can be instantiated as a QML element.
Since the default is "true" in both cases, this is somewhat confusing.

Always generate an explicit isCreatable attribute for all types to
clarify this. Also, fix the generation of isStructured as a drive-by.

Fixes: QTBUG-123089
Change-Id: I9d33d2448facc369f524ba8c68a2403ccc4c6aa5
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>