Modify tst_FlickableInterop::touchDragSliderAndFlickable to pass
We have a behavior change since Qt 5: if one finger drags a DragHandler
in a child of a Flickable, a "stray" finger that touches the Flickable
does not flick it. Time will tell whether this is a good change or not.
But it's been that way in Qt 6 so far; and in the meantime it's best
not to keep this test blacklisted because of one line.

Pick-to: 6.5
Fixes: QTBUG-123490
Task-number: QTBUG-86729
Change-Id: Iad22211b4fe102c2c1d4d7f4c7485decc0aa17a8
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
(cherry picked from commit d98f22995d9f6a401fbf77ef3c33a3fa9c316a61)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 6381928c2db5938264b2f166886bb87d5553d4e5)