qmltc: Include headers for types that appear as the argument of a list
When `qmltc` compiles a QML component, it needs to produce a set of includes
for all types that are referenced by the scope of the component so that
the produced C++ code can make use of the required C++ types.

Currently, the process that computes this include set, does not take
into consideration types that appear as the type argument to a list
property.

For example, given:

```
import QtQuick

Item {
   property list<Rectangle> foo : []
}
```

`qmltc` will not take into consideration the `Rectangle` type, and its
C++ counterpart `QQuickRectangle`.

As the generated code will need to refer to that type to compile the
`foo` property, this will produce non-compiling C++ code due to usages
of the incomplete type.

To avoid this issue, `QmltcVisitor::findCppIncludes`, which produces the
include set for the generated C++ files, was modified to produce an
include directive for types that appear as the type argument of a list
property.

A test was added to the corpus of tests for `qmltc` to test the behavior.

Fixes: QTBUG-123476
Change-Id: Ieea2895ebebe84fdc5fb92539db4e580c6fced96
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>