Remove revisioning of shadowed x/y properties in QQuickWindowQmlImpl
The properties are available in QWindow as well, and our property
resolving doesn't consider properties in base classes when a revision
excludes a property in a sublcass from being picked up.

Currently, clients who import QtQuick 2.1, and hence bring in
QQuickWindowQmlImpl instead of just QQuickWindow as the Window type,
will get:

 "Window.y" is not available in QtQuick 2.1.

The workaround is to either import QtQuick 2.0, or remove the version
from the import.

This is a partial revert of 0fdca3e40d9afed9bc745a902a7fce6329d0e2b4

Pick-to: 6.7 dev
Change-Id: I34346627e5ae12a40ddd298eea9a9e1bbc89d815
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>