CMake: Consider NO_UNSUPPORTED_PLATFORM_ERROR for non-bundle mac apps
Currently if qt6_generate_deploy_qml_app_script is called on an
executable target that does not have the MACOSX_BUNDLE property set,
it will error out with a message about not supporting non-bundle apps.

This error is shown even if NO_UNSUPPORTED_PLATFORM_ERROR is passed
as an option, which looks like an oversight.

Change the code not to show the error if NO_UNSUPPORTED_PLATFORM_ERROR
is passed. This means user projects can call
qt6_generate_deploy_qml_app_script for non-bundle apps without having
to wrap the code in a if(NOT APPLE) check, and deployment will simply
not run for that target on macOS.

[ChangeLog][Build System] The qt6_generate_deploy_qml_app_script
NO_UNSUPPORTED_PLATFORM_ERROR option will now have an effect when
calling the API on non-bundle macOS executable targets.

Pick-to: 6.7
Change-Id: Ic6b2aabf9bed46d1473134bc4306d49038013501
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>