qmltc: Avoid overwriting initial properties
When `qmltc` compiles a QML type, it allows the user to set the initial
values for the properties of the component at construction time.

Due to the order of operations that `qmltc` uses when creating a
component, the compiled type keeps track of the properties for which an
initial values was set to avoid overriding the value in later phases.

For example, a property that is bound by its definition in the original
QML file, will have that bound applied after the initial values for the
properties are set.

To avoid this issue, a type compiled by `qmltc` uses its knowledge of
which property was initialized to avoid setting the binding for the
property in the later phases.

Due to an error in the code-base for `qmltc`, certain bindings for a
property are still being set independently of the property having an
initial value or not.

In particular, one of the branches in
`QmltcCodeGenerator::generate_createBindingOnProperty` was not
conditioning the generated property binding on whether an initial value
was set.

For example, a value list property that is bound to an empty list, would
have passed by the incorrect branch, so that its value would be
overwritten even when the property was provided with an initial value.

To avoid the issue, the branch was modified to generate code that takes
into account the initialization, or lack thereof, of the property.

The test that verifies the initial values behavior was modified to have
a representative case for the issue.

Task-number: QTBUG-120700
Change-Id: Ic3f56a8774e26ab18d3fd234a1cd28e525672e36
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>