qmltc: Support setting initial values for reference list properties
When `qmltc` compiles a QML type to C++, it generates a proxy object for
the compiled type, `PropertyInitializer`, that knowns how to set the
top-level property of the component.

A user can then pass a call-back to the constructor for the compiled
type and use an instance of the generated `PropertyInitializer` to set
the initial values for the top-level properties of the component.

Currently, the generated `PropertyInitializer` would refuse to generate
a setter for properties whose type is a list of object types, impeding
the user from being able to set the initial values for properties of
this kind.

To allow setting the initial values for properties whose type is a list
of object types, `qmltc` will generate a setter for those properties
when generating the `PropertyInitializer` class for a type.

The free function `compilePropertyInitializer` in "qmltccompiler.cpp",
that is responsible for generating the `PropertyInitializer` class for a
type, was modified to produce a setter for properties whose type is a
list of object types.

The test that verifies the initial values behavior was modified to have
representative cases for the modified feature.

Fixes: QTBUG-120700
Change-Id: I4a023263ca3dfa3c77c11d1522b8be47b7958109
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>