qmltc: Ensure usage of generated setters in `PropertyInitializer`
When `qmltc` compiles a QML type to C++, it generates a proxy object for
the compiled type, `PropertyInitializer`, that knows how to set the
top-level property of the component.

A user can then pass a call-back to the constructor for the compiled
type and use an instance of the generated `PropertyInitializer` to set
the initial values for the top-level properties of the component.

For each top-level property, the generated `PropertyInitializer`
generates a setter.

Based on various parameters, the generated setter can have a different
body.

Generally, the generated setter will prefer calling back to the
component setter for the properties when that is available.

When that is not the case, the generated setter would fall-back to a
more general approach based on `QObject::setProperty`, which has
additional overhead.

This second approach will currently be chosen, when generating the
setter for an owned property of the compiled type.

While `qmltc` will generate a typed setter for those properties, on the
C++ side, at the
point where `PropertyInitializer` is compiled, the inspected owned properties,
that come from the QML side, will not have a write accessor, which in
turn makes the generated code fall-back to the more general
implementation for the setter of those properties.

To ensure that the generated setters for `PropertyInitializer` use a
write accessor when present, `compilePropertyInitializer`, which takes
care of producing the code for the generated `PropertyInitializer`, was
modified to special case the generation of the setter for properties
that are owned by the compiled type, ensuring that the generated setter
will be used to set the initial value for the property.

Change-Id: I0c34d6dd9d85a613d1e8909e94a584bf976a009f
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>