qmltc: Test failure on presence of non-top level required property
`qmltc` should fail when it processes a QML file that has a required
property that is not top-level, as it is invalid QML.

For example:

```
import QtQuick

Item {
  Item {
    required property int foo
  }
}
```

To parse a QML file, `qmltc` uses `QQmlJSImportVisitor`, which will
produce an error on encountering such a property.

`qmltc` will in turn propagate the error and fail itself, thus acting
properly when a non-top level required property is encountered in a QML
file.

A new test, `qmltc_build_failures`, is added to keep track of the
behavior and ensure that `qmltc` fails at the build-level when this form
of error is encountered.

Task-number: QTBUG-120698
Change-Id: I45f975f1258a42fd2eb66283c1d3611152278fcc
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>