qmlls: fix crash when loading a file multiple times
Fix a qmlls crash that happened due to a refactoring mishap: the helper
method getItemIfHasSomeCode always returns null, but the callers assert
that the result is not null.

Assign to value of the outer scope instead
of creating a second (unused) value variable in the inner scope.

This happens when loading an identical file twice, once in an original
environment and once in its child.

Amends 45a99df60ea4e3fddf976f862e9814fc177071d0 that did not get in 6.7
so no need to pick back.

Fixes: QTBUG-123591
Change-Id: Ie413681bb96505111c7e587b27a1b6fb819505f4
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>