CMake: Move generated qmltypes file into .qt subdirectory of build dir
So we have a single central location for most files.

[ChangeLog][Build System] Generated .qmltypes files will now be placed
into the .qt/qmltypes subdirectory of a target build directory. The
location is an implementation detail that might still change in the
future, so it should not be relied upon.

Change-Id: Ifd26e8eef0082c1f2d3ad916691c6b57aa8cb14e
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>
(cherry picked from commit 6777c19e398021b671dd6fdb8111490a9be29f7e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>