Loader: keep control over resizing
A Loader with an explicitly set size is supposed to control its child's
size. However, so far, we would have kept bindings on width and height
in place - before Qt 6.2 in any case, since 6.2 (and the change to make
Item's geometry properties bindable) only when both size and width
change at the same time. Thus, if the binding was triggered afterwards,
it would have overridden the size set by the Loader.
Fix this by always removing installed bindings.

Fixes: QTBUG-117641
Pick-to: 6.7
Change-Id: Iece4d90bacd9e16a07e5018bf11fa0f7ba9c84df
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>