dom: dot separated identifiers for all bindings
Amends a44f21f19462cc79f82080404515c4322b7728ee that forgot to set the
bindingIdentifiers for object and array bindings (it only did script
bindings).

For some reasons, this makes
tst_qmlls_modules::documentFormatting(verbatimString.qml) crash. Fix
that crash by explicitly disabling scriptexpressions on the
(scriptexpression-)unsupported AST::TemplateLiteral.

Add a const to fieldMemberExpressionForQualifiedId()'s argument to be
able to distinguish between the "input" and "output" arguments of
setBindingIdentifiers(): identifiers is supposed to be read, and
bindingPtr is supposed to be modified.

Fixes: QTBUG-123619
Change-Id: I02e164ea22416fc13202a119896ad21cc82e5d41
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>