qmlls: test deferred/generalized grouped properties in qmlls
Change propertyBindingFromReferrerScope() to accept a TypeResolver as
parameter, and use it to resolve ids for bindings to deferred
non-properties. Add a comment explaining why deferred non-properties are
ids and remove the previously existing confusing comment.

Do not discard the expression type or the expression name if the scope
was not found, and just leave the semantic scope from the returned
value out.

Add some tests for deferred grouped properties in qmlls and add some
fixes to make the tests run.

Fixes: QTBUG-123618
Change-Id: If56945acfc84b092c07760bbefe74825193c8808
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>