Add missing file locations in Import
FileLocationRegions are useful to get source location information for a
particular dom item. In semantic highlighting, we need the source
location of the token to encode it properly. This commit adds the file
location information for Import dom elements' fields using the relavant
regions. Also introduce new region so-called VersionRegion which keeps
the source location of versioning tokens.

Task-number: QTBUG-123775
Task-number: QTBUG-120000
Change-Id: Ifca472676049eb0039dfd0170f0738f9a6764cfe
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>