qmltest: add a 'modifiers' argument to mouseMove()
As it stood, TestCase::mouseMove() didn't take a 'modifier'
argument. All the other mouse functions do that, including
mouseDrag(). But even for mouseDrag, the modifier argument was
only applied to mousePress() and mouseRelease(), but not to
the in-between mouse moves.

This patch will implement support for providing a modifier
argument also to mouseMove(). This is needed in order to
test API that distinguishes a normal drag from a ctrl+drag, like
e.g SelectionRectangle.

Pick-to: 6.6 6.5
Change-Id: Idd800eda4b5b4fd4e9cda94155ca4ec36e935d13
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
(cherry picked from commit 92f39a3919375af7e0eeaa31d010fc7fde7f6c4f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>