QML Profiler: Avoid reporting scene graph data out of order
The data is recorded by different threads. Since the timestamps are
recorded in the thread context but the data is appended to a common
list, the events can arrive out of order. If that happens, insert at the
right place.

This fixes the flakiness in the tst_qqmlprofilerservice test.

Pick-to: 6.5
Change-Id: I6b3a86282c6f2e0fb8f0d41ea96c9904dba925ff
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
(cherry picked from commit 47c702ef77a9319e6bdff03da3233f92e92cda85)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>