QQmlJSImportVisitor mark required aliases as required in their scope
When `QQmlJSImportVisitor` visits a program, it marks properties as
required in their scope when it encounters them or it encounters a
`required` statement.

For aliases, an additional step is performed at the end of a visit, when
the aliases are resolved.

After an alias is resolved to a target property, is "requiredness", is
set based on the required of the target property in its scope.

Due to the specific implementation, the alias will be set to required in
its scope if the target property is required in its scope and will be
set to be non-required otherwise, independently of its previous required
state.

When we have an alias that is itself required pointing to a property
that isn't required, for example:

```
Item {
   id: self
   property int foo
   property alias alias: self.foo
   required alias
}
```

The alias will always be marked as not required, as it always inherits
the required state of the target property, losing the required status
that was applied by the "required" statement.

To respect "required" statements in relation to aliases,
`QQmlJSImportVisitor::resolveAliasesAndIds` will now inherit the
"requiredness" of the target property only when the alias is not already
marked as required in its scope.

Fixes: QTBUG-123837
Change-Id: I7b0f6958287625f3f3a5d736d4c45a3e050d4502
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>