qmldomcomments: include comment start string in the comment location
Parser doesn't take comment start strings into account. So, // or /* is
not the part of the source location. Add this via this post-processing.
The full comment region is required for semantic highlighting.

This is obviously not the correct layer for a fix. The actual fix should
go in the parser level which is tracked by QTBUG-124897.

Task-number: QTBUG-124897
Task-number: QTBUG-120000
Change-Id: I1759edf512c21809955531ba6d5dfe4ec2f7d678
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>