QQuickPaintedItem: make the antialiasing property work
The QQuickPaintedItem class was implemented with its own antialiasing
setting and [set]antialiasing() functions. Those were never reached
from qml side however, since the the parent class, QQuickItem, has an
antialiasing property, with non-virtual accessor functions of the same
names. The result was that the paint() function in QQuickPaintedItems
would always be called with a QPainter without the Antialiasing render
hint set.

Fix by dropping the redundant local antialiasing setting in
QQuickPaintedItem. Its accessor functions can presumably not be
removed yet because of bic, but their implementation is changed to
just call the corresponding accessor in QQuickitem.

Fixes: QTBUG-124170
Change-Id: I004fc59a754e30abe4390ffc8aa1015dcf493c36
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>