QQuickMenuBarPrivate: only enable native menu bars on macOS
Equal to Widgets, we should only enable native QPlatformMenuBars
on macOS. The other QPA platforms, especially Windows, contains
QPlatformMenus that are not in use by Widgets, doesn't look native,
and have bit-rotted. So we skip using them for now.

Moreover, since we cannot hide a QPlatformMenu from the menu bar,
we need to avoid creating them in the first place when they're
hidden. A follow-up patch will also detect changes to
MenuBarItem.visible at run-time.

Also note that when we don't use a native menu bar, we cannot use
native menus as drop-down menus either. Opening up a native menu
will typically result in the menu running in a separate native event
loop, which will stop mouse events from being sent to Qt. And receiving
mouse events is needed for the non-native menu bar to work. So it's
an all or nothing approach.

Change-Id: I7962b42844f2cf409e0676d755b1f928c1d16ed0
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>